requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
google-generativeai
python-dotenv
pydantic
